both types of tests now compile
