+++ rust/psibase-macros-test/services/basic-w-query/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "basicwquery"
+version.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+edition = "2021"
+publish = false
+
+[package.metadata.psibase]
+package-name = "basicwquery"
+# RPC for this service (not optional if RPC service is same as server; omission means no RPC service)
+server = "basicwquery"
+# postinstall (can be anywhere, but better to be with the service)
+# the params here just represent an action; rawData is <no args> in Fracpack terms
+# only required if service requires a postinstall/init() call
+# postinstall = [
+#     { sender = "basicwquery", service = "basicwquery", method = "init", rawData = "0000" },
+# ]
+
+[package.metadata.psibase.dependencies]
+HttpServer = "0.12.0"
+
+[dependencies]
+psibase = { path = "../../../psibase" }
+psibase_macros = { path = "../../../psibase_macros" }
+async-graphql = "7.0.7"
+serde = { version = "1.0.209", features = ["derive"] }
+mincallrecvr = { path = "../min-call-recvr" }
+
+[dev-dependencies]
+psimacrotest = { path = "../..", version = "0.12.0" }
