+++ rust/psibase-macros-test/services/min-call-recvr/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "mincallrecvr"
+version.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+edition = "2021"
+publish = false
+
+[package.metadata.psibase]
+package-name = "mincallrecvr"
+# server = "mincallrecvr"
+# postinstall = [
+#     { sender = "mincallrecvr", service = "mincallrecvr", method = "init", rawData = "0000" },
+# ]
+
+[package.metadata.psibase.dependencies]
+HttpServer = "0.12.0"
+
+[dependencies]
+psibase = { path = "../../../psibase" }
+psibase_macros = { path = "../../../psibase_macros" }
+async-graphql = "7.0.7"
+serde = { version = "1.0.209", features = ["derive"] }
+
+[dev-dependencies]
+psimacrotest = { path = "../..", version = "0.12.0" }
