+++ rust/psibase-macros-test/services/min-call-recvr/src/lib.rs
@@ -1,0 +1,8 @@
+#[psibase::service(name = "mincallrecvr")]
+mod service {
+    #[action]
+    // got an error because serde dep wasn't present?
+    fn add2(a: i32, b: i32) -> i32 {
+        a + b
+    }
+}
